requirements.txt
pytest==7.4.3
pytest-html==4.1.1
allure-pytest==2.13.2
requests==2.31.0
